Update twitter script to hide side boxes and post stats.
Hide the login and (optionally) trending side boxes that appear on
wider displays.
Hide (optionally) the comment, retweet etc. counts below posts and the
profile follower count.
Also add the media and likes tabs, and the little post interaction
drop-down to the list of things we hide because we're not logging in.